+++ src/main/java/com/patchworkmc/LambdaUtil.java
@@ -1,0 +1,6 @@
+package com.patchworkmc;
+
+public class LambdaUtil {
+	// EventBus constants
+
+}
