+++ src/main/java/com/patchworkmc/event/initialization/RegisterEventRegistrars.java
@@ -9,8 +9,7 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
 
-import com.patchworkmc.event.EventHandlerScanner;
-
+import static com.patchworkmc.Constants.*;
 public class RegisterEventRegistrars implements Consumer<MethodVisitor> {
 	private static final String EVENT_REGISTRAR_REGISTRY = "net/minecraftforge/eventbus/api/EventRegistrarRegistry";
 	private static final String EVENT_REGISTRAR_REGISTRY_INSTANCE_SIGNATURE = "Lnet/minecraftforge/eventbus/api/EventRegistrarRegistry;";
@@ -35,9 +34,9 @@ public class RegisterEventRegistrars implements Consumer<MethodVisitor> {
 			method.visitVarInsn(Opcodes.ALOAD, 1); // Push the instance to the stack (1)
 			method.visitLdcInsn(Type.getObjectType(className)); // Push the class to the stack (2)
 
-			method.visitInvokeDynamicInsn("accept", "()Ljava/util/function/Consumer;", EventHandlerScanner.METAFACTORY,
-				EventHandlerScanner.OBJECT_METHOD_TYPE, new Handle(Opcodes.H_INVOKESTATIC, className, EventHandlerScanner.REGISTER_STATIC, EventHandlerScanner.REGISTER_STATIC_DESC, false),
-				Type.getMethodType(EventHandlerScanner.REGISTER_STATIC_DESC)); // Push the lambda to the stack (3)
+			method.visitInvokeDynamicInsn("accept", "()Ljava/util/function/Consumer;", Lambdas.METAFACTORY,
+				Lambdas.OBJECT_METHOD_TYPE, new Handle(Opcodes.H_INVOKESTATIC, className, EventBus.REGISTER_STATIC, EventBus.REGISTER_STATIC_DESC, false),
+				Type.getMethodType(EventBus.REGISTER_STATIC_DESC)); // Push the lambda to the stack (3)
 			// Pop the instance for calling, and then pop the class lambda as parameters
 			method.visitMethodInsn(Opcodes.INVOKEINTERFACE, EVENT_REGISTRAR_REGISTRY, "registerStatic", "(Ljava/lang/Class;Ljava/util/function/Consumer;)V", true);
 		}
@@ -46,9 +45,9 @@ public class RegisterEventRegistrars implements Consumer<MethodVisitor> {
 			method.visitVarInsn(Opcodes.ALOAD, 1); // Push the instance to the stack (1)
 			method.visitLdcInsn(Type.getObjectType(className)); // Push the class to the stack (2)
 
-			method.visitInvokeDynamicInsn("accept", "()Ljava/util/function/BiConsumer;", EventHandlerScanner.METAFACTORY,
-				Type.getMethodType("(Ljava/lang/Object;Ljava/lang/Object;)V"), new Handle(Opcodes.H_INVOKESTATIC, className, EventHandlerScanner.REGISTER_INSTANCE, EventHandlerScanner.getRegisterInstanceDesc(className), false),
-				Type.getMethodType(EventHandlerScanner.getRegisterInstanceDesc(className))); // Push the lambda to the stack (3)
+			method.visitInvokeDynamicInsn("accept", "()Ljava/util/function/BiConsumer;", Lambdas.METAFACTORY,
+				Type.getMethodType("(Ljava/lang/Object;Ljava/lang/Object;)V"), new Handle(Opcodes.H_INVOKESTATIC, className, EventBus.REGISTER_INSTANCE, EventBus.getRegisterInstanceDesc(className), false),
+				Type.getMethodType(EventBus.getRegisterInstanceDesc(className))); // Push the lambda to the stack (3)
 			// Pop the instance for calling, and then pop the class lambda as parameters
 			method.visitMethodInsn(Opcodes.INVOKEINTERFACE, EVENT_REGISTRAR_REGISTRY, "registerInstance", "(Ljava/lang/Class;Ljava/util/function/BiConsumer;)V", true);
 		}
