+++ src/main/java/com/patchworkmc/transformer/PatchworkTransformer.java
@@ -177,10 +177,6 @@ public class PatchworkTransformer implements BiConsumer<String, byte[]> {
 		checker.onClassScanned(name, subscribeEvents, supers);
 	}
 
-	public void accept(Consumer<String> consumer) {
-
-	}
-
 	/**
 	 * Finishes the patching process.
 	 *
